chore: relocate /x/mcp proxy stack and dispatch helpers
https://linear.app/speakeasy/issue/AGE-2555/serve-mcp-endpointsmcp_servers-from-mcp-runtime-path-with-toolsets

Move-only refactor in preparation for serving mcp_endpoints/mcp_servers
from the /mcp runtime path. No behavior change: /x/mcp still resolves
endpoints and dispatches identically.

Interceptors plus buildProxy move from xmcp into remotemcp, encapsulated
behind a new remotemcp.ProxyManager whose Build method constructs a
configured *proxy.Proxy for a request. mcp.Service gains the dispatch
helpers (ResolveMCPEndpointAndServer, BuildResolvedMcpEndpointForServer,
LoadResolvedMcpEndpointBySlug, serveRemoteBackend, ServeMCPEndpoint) and
holds the ProxyManager. xmcp.Service shrinks to the OAuth route adapters
and .well-known handlers; xmcp.ServeMCP is now a thin wrapper that
forwards to mcp.Service.ServeMCPEndpoint with mcpRouteBase="x/mcp".
