+++ server/internal/remotemcp/proxymetrics_test.go
@@ -1,4 +1,4 @@
-package xmcp
+package remotemcp
 
 import (
 	"testing"
@@ -8,32 +8,32 @@ import (
 	"github.com/speakeasy-api/gram/server/internal/testenv"
 )
 
-func TestNewMetrics(t *testing.T) {
+func TestNewProxyMetrics(t *testing.T) {
 	t.Parallel()
 
 	meter := testenv.NewMeterProvider(t).Meter("test")
 	logger := testenv.NewLogger(t)
 
-	m := newMetrics(meter, logger)
+	m := NewProxyMetrics(meter, logger)
 	require.NotNil(t, m)
 	require.NotNil(t, m.mcpToolCallCounter)
 }
 
-func TestMetrics_RecordMCPToolCall_RecordsWithValidCounter(t *testing.T) {
+func TestProxyMetrics_RecordMCPToolCall_RecordsWithValidCounter(t *testing.T) {
 	t.Parallel()
 
 	meter := testenv.NewMeterProvider(t).Meter("test")
 	logger := testenv.NewLogger(t)
-	m := newMetrics(meter, logger)
+	m := NewProxyMetrics(meter, logger)
 
 	// Should not panic.
 	m.RecordMCPToolCall(t.Context(), "org-123", "https://x.example.com/x/mcp/server", "srv-abc", "search_tickets")
 }
 
-func TestMetrics_RecordMCPToolCall_NilCounterIsSafe(t *testing.T) {
+func TestProxyMetrics_RecordMCPToolCall_NilCounterIsSafe(t *testing.T) {
 	t.Parallel()
 
-	m := &metrics{mcpToolCallCounter: nil}
+	m := &ProxyMetrics{mcpToolCallCounter: nil}
 	// Should not panic when counter is nil.
 	m.RecordMCPToolCall(t.Context(), "org-123", "https://x.example.com/x/mcp/server", "srv-abc", "search_tickets")
 }
