+++ server/internal/remotemcp/resources_read_usage_tracking_interceptor.go
@@ -1,4 +1,4 @@
-package xmcp
+package remotemcp
 
 import (
 	"context"
