+++ server/internal/remotemcp/tools_call_authz_interceptor.go
@@ -1,4 +1,4 @@
-package xmcp
+package remotemcp
 
 import (
 	"context"
