+++ server/internal/remotemcp/tools_call_shadow_mcp_validate_and_strip_interceptor.go
@@ -1,4 +1,4 @@
-package xmcp
+package remotemcp
 
 import (
 	"context"
