+++ server/internal/remotemcp/tools_list_mcp_connect_filter_interceptor.go
@@ -1,4 +1,4 @@
-package xmcp
+package remotemcp
 
 import (
 	"context"
