+++ server/internal/remotemcp/tools_list_posthog_event_interceptor.go
@@ -1,4 +1,4 @@
-package xmcp
+package remotemcp
 
 import (
 	"context"
