+++ server/internal/remotemcp/tools_list_shadow_mcp_inject_interceptor.go
@@ -1,4 +1,4 @@
-package xmcp
+package remotemcp
 
 import (
 	"context"
